Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,7 +4,7 @@ int main(){
     int r,pie = 3.14;
     printf("Area: ");
     scanf("%d", &r);
-    int area = pie*(r*r);
+    float area = pie*(r*r);
     printf("%d", area);
     return 0 ;
